requirements.txt
pbr>=0.5.21,<1.0
SQLAlchemy>=0.7.8,<=0.7.99
amqplib>=0.6.1
anyjson>=0.3.3
argparse
boto>=2.4.0,!=2.13.0
eventlet>=0.13.0
Jinja2
kombu>=2.4.8
lxml>=2.3
Routes>=1.12.3
WebOb>=1.2.3,<1.3
greenlet>=0.3.2
PasteDeploy>=1.5.0
Paste
sqlalchemy-migrate>=0.7.2
netaddr
suds>=0.4
paramiko>=1.8.0
pyasn1
Babel>=1.3
iso8601>=0.1.8
jsonschema>=1.3.0,!=1.4.0
six>=1.4.1
stevedore>=0.10
websockify>=0.5.1,<0.6
oslo.config>=1.2.0